Created the basic UI scaffolding for the browser.
This is still kind of placeholder code, but it is more useful
placeholder code.  Demonstrates the basics of GTK/glade that we'll need
for this project.

When you open the demo now, there are two different browser tabs, each
running in their own universe.

Unfortunately, the browser is not quite functional yet, as things like
the url bar are only connected to placeholder signal handlers.