Moving subprocess stuff into universe.py.
This is to both separate the subprocess logic from the UI code, as well
as to abstract out the IPC stuff, so a more robust scheme can be added
later.

This should also make it easy to later abandon the gtk frontend in favor
of something more painlessly cross platform.

Currently this assumes a 1:1 relationship between universes and tabs,
but later universes may control several tabs.