Basic bidirectional interprocess communication scheme.
This scheme uses pipes to communicate with subprocesses asynchronously.
The IpcHandler class creates a thread to wait for input on the read
pipe.  When input is found, it is placed in a list, and an Event object
is set.  This allows the read and write methods to be called from the
main thread in a non-blocking fashion, something the subprocess library
does not adequately provide.

Additionally, you can optionally specify a 'signal' method, which is
called when there is new data.  This comes with the caveat that the
signal method does not run on the main thread.

I think this scheme is fine for now - currently it only supports sending
singular lines, but in the future it could possibly be updated to
support json serialization of objects too.