Replaced notebook widget with custom tabulation.
The 'notebook' widget doesn't provide satisfactory control over the
appearance and behavior of the tabs, so, this commit throws together
something custom for handling tabs instead.

A horizontal sliding pane widget it used for segmenting the body of the
browser window.  The left side contains a scroll bar viewport thing with
a VBox inside of it, which is where the tab widgets are placed.

The right side contains a VBox that contains all of the sockets for the
different universe processes.  Only one socket is visible at a time, the
rest are hidden.  It seems you can't just add and remove the sockets, or
they will blank.  Retargeting a single socket also seems to be
idiosyncratic in nature.

The tab widgets right now are just button widgets, which is fine for
now, but doesn't work well with having a close button.

I put together a mechanism for showing and hiding close buttons on the
tabs when the tab is mouse-over'd, but this causes the dimensions of the
tab to change, and can cause an infinite loop, so it is dummied out for
now until I come up with something better.

This scheme is mostly good - the only thing I want to change is to
change the left pane so that it is a tree view of some kind instead.
This way, tabs can be more intuitively grouped by universe.