Addedd universe 'parents' to tab bar.
Universes are given a human readable ID, so they count up from one.
Tabs don't need to have that, so they just use UUIDs.

I also went and added a place holder for an event to trigger when you
select a universe in the tab bar, but currently nothing happens.

Ideally this will eventually show per-universe configuration settings,
but a blank page would be fine until then.  The url bar should also be
adjusted to indicate that, and if you enter something into it, it always
should try to open it as a new tab, either in the current universe or a
new or different one.

This relates to issue #4.