Added url validation, ability to open new universes.
The URL validation is able to guess if you typed in a domain name, a
search query, or a real url.

If you type in a domain name, it will default to https://, if you type
in something it doesn't understand, it opens a wikipedia search.

This resolves issue #2.

Additionally, if you open a new url, and the domain differs from that of
the open tab, then it opens the url in a new universe/tab.  So,
"maps.google.com" and "google.com" are considered to be in the same
domain for now, but "duckduckgo.com" is not.

This relates to issue #5, though there are more things I want to do for
that one.