Refactored message passing system, better routing.
Now there is no need to have an event routing table, so this reduces
the amount of boilerplate needed.

Instead, you just call "send" with the name of the handler function, and
pass all of the arguments as keyword arguments.

This relates to issue #7, as it was needed to remove the 1:1
relationship between tabs an universes.

Docstrings need to be updated.