Added destroy universe path (broken), minor bugfix
There is a code path for tearing down an entire universe and all of its
tabs in one go, but it crashes.  This is not the same as closing all of
the tabs, because it kills the universe subprocess, and tells the tabs
to just garbage collect their UI.  This bypasses a bunch of teardown
events in the subprocess itself.

I'm not sure how future-proof this approach is, though.

Also, its broken D: the browser will crash if you use it XD.

The bugfix is catching when there is a broken pipe and gracefully ish
closing down the IpcHandler.

Relates to #7 and #14.